refactor: debounce로 0.7초간 입력 없을 시 닉네임 검증
